Updates 26 plugins, removes contact-form-7, removes wangguard

+++ wp-content/plugins/broken-link-checker/modules/containers/custom_field.php
@@ -12,32 +12,32 @@ ModuleClassName: blcPostMetaManager
 */
 
 //Note : If it ever becomes necessary to check metadata on objects other than posts, it will
-//be fairly easy to extract a more general metadata container class from blcPostMeta. 
+//be fairly easy to extract a more general metadata container class from blcPostMeta.
 
 /**
  * blcPostMeta - A link container class for post metadata (AKA custom fields).
  *
  * Due to the way metadata works, this container differs significantly from other containers :
  * 	- container_field is equal to meta name, and container_id holds the ID of the post.
- *  - There is one synch. record per post that determines the synch. state of all metadata fields of that post. 
+ *  - There is one synch. record per post that determines the synch. state of all metadata fields of that post.
  * 	- Unlinking simply deletes the meta entry in question without involving the parser.
- *  - The list of parse-able $fields is not fixed. Instead, it's initialized based on the 
- *    custom field list defined in Settings -> Link Checker. 
+ *  - The list of parse-able $fields is not fixed. Instead, it's initialized based on the
+ *    custom field list defined in Settings -> Link Checker.
  *  - The $wrapped_object is an array (and isn't really used for anything).
  * 	- update_wrapped_object() does nothing.
- * 
+ *
  * @package Broken Link Checker
  * @access public
  */
 class blcPostMeta extends blcContainer {
-	
+
 	var $meta_type = 'post';
-	
+
   /**
    * Retrieve all metadata fields of the post associated with this container.
    * The results are cached in the internal $wrapped_object variable.
-   *   
-   * @param bool $ensure_consistency 
+   *
+   * @param bool $ensure_consistency
    * @return object The wrapped object.
    */
 	function get_wrapped_object($ensure_consistency = false){
@@ -45,33 +45,33 @@ class blcPostMeta extends blcContainer {
 			$this->wrapped_object = get_metadata($this->meta_type, $this->container_id);
 		}
 		return $this->wrapped_object;
-	}	
-	
+	}
+
 	function update_wrapped_object(){
 		trigger_error('Function blcPostMeta::update_wrapped_object() does nothing and should not be used.', E_USER_WARNING);
 	}
-	
+
   /**
    * Get the value of the specified metadata field of the object wrapped by this container.
-   * 
+   *
    * @access protected
    *
-   * @param string $field Field name. If omitted, the value of the default field will be returned. 
+   * @param string $field Field name. If omitted, the value of the default field will be returned.
    * @return array
    */
 	function get_field($field = ''){
 		$get_only_first_field = ($this->fields[$field] !== 'metadata');
 		return get_metadata($this->meta_type, $this->container_id, $field, $get_only_first_field);
 	}
-	
+
   /**
-   * Update the value of the specified metadata field of the object wrapped by this container. 
+   * Update the value of the specified metadata field of the object wrapped by this container.
    *
    * @access protected
    *
    * @param string $field Meta name.
-   * @param string $new_value New meta value. 
-   * @param string $old_value old meta value.     
+   * @param string $new_value New meta value.
+   * @param string $old_value old meta value.
    * @return bool|WP_Error True on success, an error object if something went wrong.
    */
 	function update_field($field, $new_value, $old_value = ''){
@@ -82,9 +82,9 @@ class blcPostMeta extends blcContainer {
 			return new WP_Error(
 				'metadata_update_failed',
 				sprintf(
-					__("Failed to update the meta field '%s' on %s [%d]", 'broken-link-checker'), 
-					$field, 
-					$this->meta_type, 
+					__("Failed to update the meta field '%s' on %s [%d]", 'broken-link-checker'),
+					$field,
+					$this->meta_type,
 					$this->container_id
 				)
 			);
@@ -112,9 +112,9 @@ class blcPostMeta extends blcContainer {
 			return new WP_Error(
 				'metadata_delete_failed',
 				sprintf(
-					__("Failed to delete the meta field '%s' on %s [%d]", 'broken-link-checker'), 
+					__("Failed to delete the meta field '%s' on %s [%d]", 'broken-link-checker'),
 					$field_name,
-					$this->meta_type, 
+					$this->meta_type,
 					$this->container_id
 				)
 			);
@@ -138,7 +138,7 @@ class blcPostMeta extends blcContainer {
 			'Editing %s[%d]:%s - %s to %s',
 			$this->container_type,
 			$this->container_id,
-			$field_name, 
+			$field_name,
 			$old_url,
 			$new_url
 		));
@@ -147,36 +147,36 @@ class blcPostMeta extends blcContainer {
 		if ( $this->fields[$field_name] !== 'metadata' ) {
 			return parent::edit_link($field_name, $parser, $new_url, $old_url, $old_raw_url, $new_text);
 		}
-		
+
 		if ( empty($old_raw_url) ){
 			$old_raw_url = $old_url;
 		}
-		
+
 		//Get the current values of the field that needs to be edited.
 		//The default metadata parser ignores them, but we're still going
-		//to set this argument to a valid value in case someone writes a 
+		//to set this argument to a valid value in case someone writes a
 		//custom meta parser that needs it.
 		$old_value = $this->get_field($field_name);
-		
+
 		//Get the new field value (a string).
 		$edit_result = $parser->edit($old_value, $new_url, $old_url, $old_raw_url);
 		if ( is_wp_error($edit_result) ){
 			return $edit_result;
 		}
-		
+
 		//Update the field with the new value returned by the parser.
 		//Notice how $old_raw_url is used instead of $old_value. $old_raw_url contains the entire old
 		//value of the metadata field (see blcMetadataParser::parse()) and thus can be used to
-		//differentiate between multiple meta fields with identical names. 
+		//differentiate between multiple meta fields with identical names.
 		$update_result = $this->update_field( $field_name, $edit_result['content'], $old_raw_url );
 		if ( is_wp_error($update_result) ){
 			return $update_result;
 		}
-		
+
 		//Return the new "raw" URL.
 		return $edit_result['raw_url'];
 	}
-	
+
   /**
    * Get the default link text to use for links found in a specific container field.
    *
@@ -189,7 +189,7 @@ class blcPostMeta extends blcContainer {
 		//the user a clue where to look if they want to find/modify the field.
 		return $field;
 	}
-	
+
 	function ui_get_source($container_field = '', $context = 'display'){
 		if ( !post_type_exists(get_post_type($this->container_id)) ) {
 			//Error: Invalid post type. The user probably removed a CPT without removing the actual posts.
@@ -216,10 +216,10 @@ class blcPostMeta extends blcContainer {
 			esc_attr(__('Edit this post')),
 			get_the_title($this->container_id)
 		);
-		
+
 		return $post_html;
 	}
-	
+
 	function ui_get_action_links($container_field){
 		$actions = array();
 		if ( !post_type_exists(get_post_type($this->container_id)) ) {
@@ -228,7 +228,7 @@ class blcPostMeta extends blcContainer {
 
 		if ( current_user_can('edit_post', $this->container_id) ) {
 			$actions['edit'] = '<span class="edit"><a href="' . $this->get_edit_url() . '" title="' . esc_attr(__('Edit this item')) . '">' . __('Edit') . '</a>';
-			
+
 			if ( $this->current_user_can_delete() ){
 				if ( $this->can_be_trashed() ) {
 					$actions['trash'] = sprintf(
@@ -248,48 +248,48 @@ class blcPostMeta extends blcContainer {
 			}
 		}
 		$actions['view'] = '<span class="view"><a href="' . esc_url(get_permalink($this->container_id)) . '" title="' . esc_attr(sprintf(__('View "%s"', 'broken-link-checker'), get_the_title($this->container_id))) . '" rel="permalink">' . __('View') . '</a>';
-		
+
 		return $actions;
 	}
-	
+
   /**
-   * Get edit URL for this container. Returns the URL of the Dashboard page where the item 
+   * Get edit URL for this container. Returns the URL of the Dashboard page where the item
    * associated with this container can be edited.
    *
-   * @access protected   
+   * @access protected
    *
    * @return string
    */
 	function get_edit_url(){
 		/*
-		The below is a near-exact copy of the get_post_edit_link() function.  
-		Unfortunately we can't just call that function because it has a hardcoded 
-		caps-check which fails when called from the email notification script 
+		The below is a near-exact copy of the get_post_edit_link() function.
+		Unfortunately we can't just call that function because it has a hardcoded
+		caps-check which fails when called from the email notification script
 		executed by Cron.
-		*/ 
-		
+		*/
+
 		if ( !($post = get_post( $this->container_id )) ){
 			return '';
 		}
-		
+
 		$context = 'display';
-		
+
 		//WP 3.0
 		if ( 'display' == $context )
 			$action = '&amp;action=edit';
 		else
 			$action = '&action=edit';
-	
+
 		$post_type_object = get_post_type_object( $post->post_type );
 		if ( !$post_type_object ){
 			return '';
 		}
-	
+
 		return apply_filters( 'get_edit_post_link', admin_url( sprintf($post_type_object->_edit_link . $action, $post->ID) ), $post->ID, $context );
 	}
-	
+
   /**
-   * Get the base URL of the container. For custom fields, the base URL is the permalink of 
+   * Get the base URL of the container. For custom fields, the base URL is the permalink of
    * the post that the field is attached to.
    *
    * @return string
@@ -297,7 +297,7 @@ class blcPostMeta extends blcContainer {
 	function base_url(){
 		return get_permalink($this->container_id);
 	}
-	
+
   /**
    * Delete or trash the post corresponding to this container. If trash is enabled,
    * will always move the post to the trash instead of deleting.
@@ -322,10 +322,10 @@ class blcPostMeta extends blcContainer {
 			}
 		}
 	}
-	
+
 	/**
 	 * Move the post corresponding to this custom field to the Trash.
-	 * 
+	 *
 	 * @return bool|WP_Error
 	 */
 	function trash_wrapped_object(){
@@ -339,13 +339,13 @@ class blcPostMeta extends blcContainer {
 				)
 			);
 		}
-		
+
 		$post = &get_post($this->container_id);
 		if ( $post->post_status == 'trash' ){
 			//Prevent conflicts between post and custom field containers trying to trash the same post.
 			return true;
 		}
-		
+
 		if ( wp_trash_post($this->container_id) ){
 			return true;
 		} else {
@@ -359,13 +359,13 @@ class blcPostMeta extends blcContainer {
 			);
 		}
 	}
-	
+
 	function current_user_can_delete(){
 		$post = get_post($this->container_id);
 		$post_type_object = get_post_type_object($post->post_type);
 		return current_user_can( $post_type_object->cap->delete_post, $this->container_id );
 	}
-	
+
 	function can_be_trashed(){
 		return defined('EMPTY_TRASH_DAYS') && EMPTY_TRASH_DAYS;
 	}
@@ -375,7 +375,7 @@ class blcPostMetaManager extends blcContainerManager {
 	var $container_class_name = 'blcPostMeta';
 	var $meta_type = 'post';
 	protected $selected_fields = array();
-	
+
 	function init(){
 		parent::init();
 
@@ -395,7 +395,7 @@ class blcPostMetaManager extends blcContainerManager {
 				}
 			}
 		}
-		
+
 		//Intercept 2.9+ style metadata modification actions
 		add_action( "added_{$this->meta_type}_meta", array($this, 'meta_modified'), 10, 4 );
 		add_action( "updated_{$this->meta_type}_meta", array($this, 'meta_modified'), 10, 4 );
@@ -404,63 +404,63 @@ class blcPostMetaManager extends blcContainerManager {
 		//When a post is deleted, also delete the custom field container associated with it.
 		add_action('delete_post', array($this,'post_deleted'));
         add_action('trash_post', array($this,'post_deleted'));
-        
+
         //Re-parse custom fields when a post is restored from trash
         add_action('untrashed_post', array($this,'post_untrashed'));
 	}
 
-	
+
 	/**
 	 * Get a list of parseable fields.
-	 * 
+	 *
 	 * @return array
 	 */
 	function get_parseable_fields(){
 		return $this->selected_fields;
 	}
-	
+
   /**
    * Instantiate multiple containers of the container type managed by this class.
    *
    * @param array $containers Array of assoc. arrays containing container data.
    * @param string $purpose An optional code indicating how the retrieved containers will be used.
-   * @param bool $load_wrapped_objects Preload wrapped objects regardless of purpose. 
-   * 
+   * @param bool $load_wrapped_objects Preload wrapped objects regardless of purpose.
+   *
    * @return array of blcPostMeta indexed by "container_type|container_id"
    */
 	function get_containers($containers, $purpose = '', $load_wrapped_objects = false){
 		$containers = $this->make_containers($containers);
-		
+
 		/*
 		When links from custom fields are displayed in Tools -> Broken Links,
 		each one also shows the title of the post that the custom field(s)
 		belong to. Thus it makes sense to pre-cache the posts beforehand - it's
 		faster to load them all at once than to make a separate query for each
 		one later.
-		
+
 		So make a list of involved post IDs and load them.
-		
-		Calling get_posts() will automatically populate the post cache, so we 
+
+		Calling get_posts() will automatically populate the post cache, so we
 		don't need to actually store the results anywhere in the container object().
-		*/    
+		*/
 		$preload = $load_wrapped_objects || in_array($purpose, array(BLC_FOR_DISPLAY));
 		if ( $preload ){
 			$post_ids = array();
 			foreach($containers as $container){
 				$post_ids[] = $container->container_id;
 			}
-			
+
 			$args = array('include' => implode(',', $post_ids));
 			get_posts($args);
 		}
-		
+
 		return $containers;
 	}
-	
+
   /**
    * Create or update synchronization records for all containers managed by this class.
    *
-   * @param bool $forced If true, assume that all synch. records are gone and will need to be recreated from scratch. 
+   * @param bool $forced If true, assume that all synch. records are gone and will need to be recreated from scratch.
    * @return void
    */
 	function resynch($forced = false){
@@ -478,7 +478,7 @@ class blcPostMetaManager extends blcContainerManager {
 		$escaped_post_types = "'" . implode("', '", array_map('esc_sql', $post_types)) . "'";
 
 		if ( $forced ){
-			//Create new synchronization records for all posts. 
+			//Create new synchronization records for all posts.
 			$blclog->log('...... Creating synch records for all custom fields on ' . $escaped_post_types);
 			$start = microtime(true);
 			$q = "INSERT INTO {$wpdb->prefix}blc_synch(container_id, container_type, synched)
@@ -501,7 +501,7 @@ class blcPostMetaManager extends blcContainerManager {
 					 synch.container_type = '{$this->container_type}' AND posts.ID IS NULL";
 			$wpdb->query( $q );
 			$blclog->log(sprintf('...... %d rows deleted in %.3f seconds', $wpdb->rows_affected, microtime(true) - $start));
- 			
+
 			//Remove the 'synched' flag from all posts that have been updated
 			//since the last time they were parsed/synchronized.
 			$blclog->log('...... Marking custom fields on changed posts as unsynched');
@@ -515,7 +515,7 @@ class blcPostMetaManager extends blcContainerManager {
 					synch.last_synch < posts.post_modified";
 			$wpdb->query( $q );
 			$blclog->log(sprintf('...... %d rows updated in %.3f seconds', $wpdb->rows_affected, microtime(true) - $start));
-			
+
 			//Create synch. records for posts that don't have them.
 			$blclog->log('...... Creating custom field synch records for new ' . $escaped_post_types);
 			$start = microtime(true);
@@ -532,7 +532,7 @@ class blcPostMetaManager extends blcContainerManager {
 			$blclog->log(sprintf('...... %d rows inserted in %.3f seconds', $wpdb->rows_affected, microtime(true) - $start));
 		}
 	}
-	
+
   /**
    * Mark custom fields as unsynched when they're modified or deleted.
    *
@@ -544,27 +544,27 @@ class blcPostMetaManager extends blcContainerManager {
    */
 	function meta_modified($meta_id, $object_id = 0, $meta_key= '', $meta_value = ''){
 		global $wpdb; /** @var wpdb $wpdb */
-		
-		//If object_id isn't specified then the hook was probably called from the 
+
+		//If object_id isn't specified then the hook was probably called from the
 		//stupidly inconsistent delete_meta() function in /wp-admin/includes/post.php.
 		if ( empty($object_id) ){
 			//We must manually retrieve object_id and meta_key from the DB.
 			if ( is_array($meta_id) ){
 				$meta_id = array_shift($meta_id);
 			}
-			
+
 			$meta = $wpdb->get_row( $wpdb->prepare("SELECT * FROM $wpdb->postmeta WHERE meta_id = %d", $meta_id), ARRAY_A );
 			if ( empty($meta) ){
 				return;
 			}
-			
+
 			$object_id = $meta['post_id'];
 			$meta_key = $meta['meta_key'];
 		}
-		
-		
-		//Metadata changes only matter to us if the modified key 
-		//is one that the user wants checked. 
+
+
+		//Metadata changes only matter to us if the modified key
+		//is one that the user wants checked.
 		if ( empty($this->selected_fields) ){
 			return;
 		}
@@ -581,7 +581,7 @@ class blcPostMetaManager extends blcContainerManager {
 		$container = blcContainerHelper::get_container( array($this->container_type, intval($object_id)) );
 		$container->mark_as_unsynched();
 	}
-	
+
   /**
    * Delete custom field synch. records when the post that they belong to is deleted.
    *
@@ -590,14 +590,17 @@ class blcPostMetaManager extends blcContainerManager {
    */
 	function post_deleted($post_id){
 		//Get the associated container object
-		$container = blcContainerHelper::get_container( array($this->container_type, intval($post_id)) );
-		//Delete it
-		$container->delete();
-		//Clean up any dangling links
-		blc_cleanup_links();
-	}
-	
-  /**
+
+        $container = blcContainerHelper::get_container(array($this->container_type, intval($post_id)));
+        if ($container != null) {
+            //Delete it
+            $container->delete();
+            //Clean up any dangling links
+            blc_cleanup_links();
+        }
+    }
+
+    /**
    * When a post is restored, mark all of its custom fields as unparsed.
    * Called via the 'untrashed_post' action.
    *
@@ -609,11 +612,11 @@ class blcPostMetaManager extends blcContainerManager {
 		$container = blcContainerHelper::get_container( array($this->container_type, intval($post_id)) );
 		$container->mark_as_unsynched();
 	}
-	
+
   /**
    * Get the message to display after $n posts have been deleted.
    *
-   * @uses blcAnyPostContainerManager::ui_bulk_delete_message() 
+   * @uses blcAnyPostContainerManager::ui_bulk_delete_message()
    *
    * @param int $n Number of deleted posts.
    * @return string A delete confirmation message, e.g. "5 posts were moved to the trash"
@@ -621,7 +624,7 @@ class blcPostMetaManager extends blcContainerManager {
 	function ui_bulk_delete_message($n){
 		return blcAnyPostContainerManager::ui_bulk_delete_message($n);
 	}
-	
+
   /**
    * Get the message to display after $n posts have been trashed.
    *
