+++ wp-content/plugins/email-log/uninstall.php
@@ -1,50 +1,71 @@
 <?php
 /**
- * Uninstall Email Log plugin
+ * Uninstall page for Email Log Plugin to clean up all plugin data.
  *
- * @package     Email Log
- * @subpackage  Uninstall
- * @author      Sudar
-*/
-// uninstall page for Email Log Plugin to clean up db.
-if( !defined( 'ABSPATH' ) && !defined( 'WP_UNINSTALL_PLUGIN' ) )
-    exit();
+ * This file is named uninstall.php since WordPress requires that name.
+ */
+
+// exit if WordPress is not uninstalling the plugin.
+if ( ! defined( 'ABSPATH' ) && ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	exit();
+}
 
 if ( is_multisite() ) {
-    global $wpdb;
+	// Note: if there are more than 10,000 blogs or
+	// if `wp_is_large_network` filter is set, then this may fail.
+	$sites = wp_get_sites();
 
-    $original_blog_id = get_current_blog_id();
-
-    $blog_ids = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
-
-    foreach ( $blog_ids as $blog_id ) {
-        switch_to_blog( $blog_id );
-        email_log_delete_table();
-    }
-
-    switch_to_blog( $original_blog_id );
-
+	foreach ( $sites as $site ) {
+		switch_to_blog( $site['blog_id'] );
+		email_log_delete_db_data();
+		restore_current_blog();
+	}
 } else {
-    email_log_delete_table();
+	email_log_delete_db_data();
 }
 
 /**
- * Delete email log table and db option
+ * Delete all email log data from db.
+ *
+ * The data include email log table, options, capability and add-on license data.
  *
  * @since 1.7
  *
  * @global object $wpdb
  */
-function email_log_delete_table() {
-    global $wpdb;
-    $table_name = $wpdb->prefix . "email_log"; // This is hardcoded on purpose
+function email_log_delete_db_data() {
+	global $wpdb;
 
-    if( $wpdb->get_var( "SHOW TABLES LIKE '{$table_name}'" ) == $table_name ) {
-        // If table is present, drop it
-        $wpdb->query( "DROP TABLE $table_name" );
-    }
+	$remove_data_on_uninstall = false;
 
-    // Delete the option
-    delete_option('email-log-db');
+	$option = get_option( 'email-log-core' );
+	if ( is_array( $option ) && array_key_exists( 'remove_on_uninstall', $option ) &&
+	     'true' === strtolower( $option['remove_on_uninstall'] ) ) {
+
+		$remove_data_on_uninstall = true;
+	}
+
+	// This is hardcoded on purpose, since the entire plugin is not loaded during uninstall.
+	$table_name = $wpdb->prefix . 'email_log';
+
+	if ( $remove_data_on_uninstall ) {
+		if ( $wpdb->get_var( "SHOW TABLES LIKE '{$table_name}'" ) == $table_name ) {
+			$wpdb->query( "DROP TABLE $table_name" );
+		}
+
+		delete_option( 'email-log-db' );
+		delete_option( 'email-log-core' );
+
+		$roles = get_editable_roles();
+		foreach ( $roles as $role_name => $role_obj ) {
+			$role = get_role( $role_name );
+
+			if ( ! is_null( $role ) ) {
+				$role->remove_cap( 'manage_email_logs' );
+			}
+		}
+
+		delete_option( 'el_bundle_license' );
+		$wpdb->query( "DELETE FROM {$wpdb->options} WHERE option_name LIKE 'el_license_%'" );
+	}
 }
-?>
